add: implement optimization test functions for benchmarking
